pyproject.toml
[tool.poetry]
name = "hermes"
version = "0.1.0"
description = "Universal data pipeline for preparing graph databases"
authors = ["Todd"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pydantic = "^2.5.0"
numpy = "^1.26.0"
pandas = "^2.1.0"
python-arango = "^7.6.0"
PyPDF2 = "^3.0.0"
pypdf = "^3.17.0"
docling = "^2.0.0"
pdf2image = "^1.16.0"
jina = "^3.20.0"
torch = "^2.1.0"
sentence-transformers = "^2.2.2"
tqdm = "^4.66.0"
pyyaml = "^6.0.1"
python-dotenv = "^1.0.0"
click = "^8.1.7"
rich = "^13.7.0"
dspy = "^2.6.27"
pdfplumber = "^0.11.7"
faiss-cpu = "^1.11.0.post1"
peft = "^0.16.0"
vllm = "^0.9.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.11.0"
mypy = "^1.7.0"
ruff = "^0.1.6"
ipykernel = "^6.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true